feat(table): adopt native cell selection on upgraded gpui-component
Upgrade gpui/gpui-component to git main (unreleased 0.5.2) — the only line
with native cell selection — and rework the table on top of it:

- fix upgrade breakage: owned Column, Table -> DataTable, infallible
  AsyncApp::update, Dialog confirm -> button_props + show_cancel,
  Menu.disabled, Application::new -> gpui_platform::application()
- drop the custom single-cell selection layer; a TableEvent bridge in
  TablePanel caches the native selection into a delegate cursor, bounces
  clicks off the pinned # column, starts edits on DoubleClickedCell, and
  re-emits TableChanged so readers track keyboard navigation too
- enable cell/row/column selection and movable + resizable columns;
  persist column order/widths per project and restore on load
- Details panel: skeleton image placeholder + DescriptionList of the
  selected row; re-binds when the table is rebuilt (fixes the panel being
  stuck on "No selection"), same for the status-bar cell readout
- open windows synchronously: gpui main quits on an empty window list, so
  async opens killed the app between window transitions

Refs: devnull03/qrate#43
Task: https://www.notion.so/39c21d32b13b81448a64f12b2dd7afed

+++ crates/settings/src/lib.rs
@@ -233,7 +233,7 @@ pub struct MainWindowBounds {
     pub height: f32,
     /// `PlatformDisplay::id` as `u32`; matched at startup against [`App::displays`].
     #[serde(default)]
-    pub display_id: Option<u32>,
+    pub display_id: Option<u64>,
 }
 
 impl MainWindowBounds {
@@ -242,7 +242,7 @@ impl MainWindowBounds {
         Self {
             width: b.size.width.into(),
             height: b.size.height.into(),
-            display_id: window.display(cx).map(|d| u32::from(d.id())),
+            display_id: window.display(cx).map(|d| u64::from(d.id())),
         }
     }
 
@@ -258,7 +258,7 @@ impl MainWindowBounds {
         let display = bounds.and_then(|b| b.display_id).and_then(|raw| {
             cx.displays()
                 .into_iter()
-                .find(|d| u32::from(d.id()) == raw)
+                .find(|d| u64::from(d.id()) == raw)
                 .map(|d| d.id())
         });
         const MIN_W: f32 = 400.0;
